+++ posthog/hogql/property.py
@@ -207,21 +207,13 @@ def property_to_expr(property: Union[BaseModel, PropertyGroup, Property, dict, l
         cohort = Cohort.objects.get(team=team, id=property.value)
         return ast.CompareOperation(
             left=ast.Field(chain=["person_id"]),
-            op=ast.CompareOperationOp.In,
-            right=cohort_subquery(cohort.pk, cohort.is_static),
+            op=ast.CompareOperationOp.InCohort,
+            right=ast.Constant(value=cohort.pk),
         )
 
     # TODO: Add support for these types "group", "recording", "behavioral", and "session" types
 
     raise NotImplementedException(f"property_to_expr not implemented for filter type {type(property).__name__}")
-
-
-def cohort_subquery(cohort_id, is_static) -> ast.Expr:
-    if is_static:
-        sql = "(SELECT person_id FROM static_cohort_people WHERE cohort_id = {cohort_id})"
-    else:
-        sql = "(SELECT person_id FROM raw_cohort_people WHERE cohort_id = {cohort_id} GROUP BY person_id, cohort_id, version HAVING sum(sign) > 0)"
-    return parse_expr(sql, {"cohort_id": ast.Constant(value=cohort_id)}, start=None)  # clear the source start position
 
 
 def action_to_expr(action: Action) -> ast.Expr:
